// src/app/pages/models/supportdata.model.ts
export interface supportData {
            id: string;
            questionid: string;                
            qcontent: string; 
            rcontent: string; 
            qstatus: string; 
            subject: string; 
            agentid: string; 
       name: string;
       googleid: string;
       loginemail: string;
       loginphone: string;

       updatedate: string;
       country: string;
};